parser: Implement Deserialize & Serialize for types
Originally I tried to implement these behind cfg(test) but soon realized
that doesn't actually take effect in integration tests which is totally
fair. The workaround to that is to have a feature but that would require
us to enable the feature explicitly while running the tests (or set it
as a default feature which slightly defeats the purpose of this shtick).
However, even that is not exactly very nice today because you can't mark
a feature as required for a particular integration test, see:
https://github.com/rust-lang/cargo/issues/2911

Anyway this whole dance is not really necessary, there's no reason to
intentionally avoid having seralization in the parser, unless a reason
to avoid having it comes up (maybe slower compile times due to proc-
macro but very unlikely), then this decision can be revisited