generator: Create a type-elided HTML generator
This is an early version of it without many of the expected features,
obviously.

I looked into using swc_html_ast and other libraries. A lot of the
libraries try to have type-safe HTML too much which is not ideal for our
usecase where the language itself will check the HTML at runtime and the
compile time checks would be nothing more than a nuance.
While swc_html_ast feels perfect, it has a lot of opinionated decisions
that would affect Effectful too much, for example it uses a specific
library for interning strings and _requires_ that all strings be
interned with it with no option to opt out of that, which is really
not ideal. Sadly this means we probably won't be able to make use of
swc_ecma_ast as well, but maybe that's for the better, we don't really
need access to _all_ of JS after all, for example we will never compile
into classes and co